+++ lib/src/languages/typescript/ts/typescript_grammar_lexer.dart
@@ -73,6 +73,12 @@ abstract class TypeScriptGrammarLexer extends BaseGrammarLexer {
 
   Parser whileToken() => ref1(token, 'while');
 
+  Parser tryToken() => ref1(token, 'try');
+
+  Parser finallyToken() => ref1(token, 'finally');
+
+  Parser throwToken() => ref1(token, 'throw');
+
   // TypeScript-specific keywords.
   Parser abstractToken() => ref1(token, 'abstract');